feat: combine bot subreport into API report
